+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.12
 * Fixed query parameter issue not handled properly (by https://github.com/shreyas18jan)
+* Removed shake dependency and added sensors dependency. Shake will be detected with sensors.
 
 ## 0.1.11
 * Updated dependencies